+++ python_jobs/src/hbnmigration/utility_functions/__init__.py
@@ -1,5 +1,6 @@
 """Utility functions."""
 
+from .cache import DataCache, get_cache_dir, get_recent_time_window, YESTERDAY
 from .custom import (
     create_tempory_file,
     execute_vars_file,
@@ -19,7 +20,6 @@ from .custom import (
     read_vars_file_as_module,
     redcap_api_push,
     today,
-    yesterday,
     yesterday_or_more_recent,
 )
 from .datatypes import (
@@ -47,6 +47,7 @@ from .secrets import ImportWithFallback
 from .typescript import tsx
 
 __all__ = [
+    "YESTERDAY",
     "ApiProtocol",
     "ApiProtocols",
     "CliOptions",
@@ -57,6 +58,7 @@ __all__ = [
     "CuriousDecryptedAnswer",
     "CuriousEncryptedAnswer",
     "CuriousId",
+    "DataCache",
     "Endpoints",
     "FieldDescriptor",
     "ImportWithFallback",
@@ -74,7 +76,9 @@ __all__ = [
     "fetch_api_data2",
     "fetch_api_data3",
     "flat_to_eav",
+    "get_cache_dir",
     "get_mindlogger_token",
+    "get_recent_time_window",
     "get_redcap_event_names",
     "initialize_logging",
     "new_curious_account",
@@ -87,6 +91,5 @@ __all__ = [
     "setup_tsv_logger",
     "today",
     "tsx",
-    "yesterday",
     "yesterday_or_more_recent",
 ]