+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved project status to envrironment variable `$HBNMIGRATION_PROJECT_STATUS` with default `"prod"`
 - Log missing fields but push acceptable data in those cases
 - Log invalid categories separately
+- Optimize from day-by-day to minute-by-minute
 
 ### Fixed